dump helper functions in different modules
